python/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
aiomysql
jinja2
python-multipart
openai
tiktoken